fix overflow to translate XLM into stroops

+++ src/main/CommandHandler.cpp
@@ -115,7 +115,7 @@ void CommandHandler::testTx(std::string const& params, std::string& retStr)
         else fromKey = getAccount(from->second.c_str());
 
         int64_t txfee = mApp.getLedgerManager().getTxFee();
-        uint64_t paymentAmount = stoi(amount->second)*1000000;
+        uint64_t paymentAmount = uint64_t(stoi(amount->second))*1000000ULL;
 
         SequenceNumber fromSeq = getSeq(fromKey, mApp) + 1;
